backend/requirements.txt
asgiref==3.8.1
pytz==2024.2
sqlparse==0.5.2
Django==3.2.16
pillow==11.0.0
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0